+++ src/packerlicious/provisioner.py
@@ -1,11 +1,62 @@
-from . import BasePackerObject, EnvVar, TemplateVar
 import validator
+import warnings
+
+from . import BasePackerObject, EnvVar, TemplateVar
 
 
 class PackerProvisioner(BasePackerObject):
 
     def __init__(self, title=None, **kwargs):
         super(PackerProvisioner, self).__init__(title, **kwargs)
+
+
+class AnsibleLocal(PackerProvisioner):
+    """
+    Ansible Local Provisioner
+    https://www.packer.io/docs/provisioners/ansible-local.html
+    TODO inventory_file validation of --limit
+    """
+    resource_type = "ansible-local"
+
+    props = {
+        'playbook_file': (basestring, True),
+        'command': (basestring, False),
+        'extra_arguments': ([basestring], False),
+        'inventory_groups': (basestring, False),
+        'inventory_file': (basestring, False),
+        'playbook_dir': (basestring, False),
+        'playbook_paths': ([basestring], False),
+        'galaxy_file': (basestring, False),
+        'group_vars': (basestring, False),
+        'host_vars': (basestring, False),
+        'role_paths': ([basestring], False),
+        'staging_directory': (basestring, False),
+    }
+
+
+class Ansible(PackerProvisioner):
+    """
+    Ansible Provisioner
+    https://www.packer.io/docs/provisioners/ansible.html
+    """
+    resource_type = "ansible"
+
+    props = {
+        'playbook_file': (basestring, True),
+        'ansible_env_vars': ([basestring], False),
+        'command': (basestring, False),
+        'empty_groups': ([basestring], False),
+        'extra_arguments': ([basestring], False),
+        'groups': ([basestring], False),
+        'host_alias': (basestring, False),
+        'inventory_directory': (basestring, False),
+        'local_port': (basestring, False),
+        'sftp_command': (basestring, False),
+        'skip_version_check': (validator.boolean, False),
+        'ssh_host_key_file': (basestring, False),
+        'ssh_authorized_key_file': (basestring, False),
+        'user': (basestring, False),
+    }
 
 
 def valid_file_direction(x):
@@ -30,6 +81,43 @@ class File(PackerProvisioner):
     # File Provisioner constants
     Upload = "upload"
     Download = "download"
+
+
+class SaltMasterless(PackerProvisioner):
+    """
+    Salt Masterless Provisioner
+    https://www.packer.io/docs/provisioners/salt-masterless.html
+    """
+    resource_type = "salt-masterless"
+
+    props = {
+        'local_state_tree': (basestring, True),
+        'bootstrap_args': (basestring, False),
+        'disable_sudo': (validator.boolean, False),
+        'remote_pillar_roots': (basestring, False),
+        'remote_state_tree': (basestring, False),
+        'local_pillar_roots': (basestring, False),
+        'custom_state': (basestring, False),
+        'minion_config': (basestring, False),
+        'grains_file': (basestring, False),
+        'skip_bootstrap': (validator.boolean, False),
+        'temp_config_dir': (basestring, False),
+        'no_exit_on_failure': (validator.boolean, False),
+        'log_level': (basestring, False),
+        'salt_call_args': (basestring, False),
+        'salt_bin_dir': (basestring, False),
+    }
+
+    def validate(self):
+        conds = [
+            'remote_pillar_roots',
+            'remote_state_tree',
+            ]
+
+        if 'minion_config' in self.properties and validator.count(self.properties, conds) > 0:
+            # TODO should this just bee an error?
+            warnings.warn("'minion_config' is present, 'remote_pillar_roots' and 'remote_state_tree' will be ignored.")
+
 
 
 class Shell(PackerProvisioner):